metrics-pipeline/metrics_pipeline/requirements.txt
dagster
dagster-webserver
dagster-dbt
polars
loguru
pyarrow
connectorx
dagster-gcp
dagster-gcp-pandas
upstash_redis
dbt-core
dbt-bigquery
dbt-postgres